Update raid.py
I can understand that, yea. How about an informative command then? That way people would join the support server and come to the raid.
I really like the raids, but only a fraction of the players actually know about them, so this command could give some information. But if you don't like that idea either, you can ignore this, of course.