paginate classes
I have no experience with the paginator, actually. No idea If this would then also send the message with the embeds as well, but I hope everything is correct like this. Sorry for letting you wait.
If anything is up with the descriptions that I should change, do let me know.

Also, you didn't comment on the change to `$hunt`. Do you agree with that change?