Update bot presence?
I'm not sure if this will work, but maybe, because the bot is a little funky on startup, this *could* fix this issue?
Definitely, the huge amount of servers and users the bot connects to plays into this.

Especially the part where you mentioned "Has worked always" makes me think this might be the issue.

I also thought it might have something to do with the bot not being ready for a long time. Last time I saw you restart the bot, when you checked `bot.is_ready()` repeatedly, it took about ten minutes for it to finally be ready.

Another possibility would be to remove the line 93.

Of course there's no easy way to test this, maybe it would be possible to test using the beta bot?