fix gamemaster command access logic, this time for real
Previously, if a user was a gamemaster, but not an owner, they would be stopped by the second check, due to the game master cog also being part of the owner_exts. The attempted owner bypass has been implemented also, so that bot owners who are not gamemasters can access the help for these commands regardless.